src/scheduler/pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
description = "Pipeline scheduler service for NYC TLC analytics"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "psycopg2-binary>=2.9.0",
    "prefect>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "testcontainers>=4.9.0",
    "ruff>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
